requirements.txt
keyring
bcrypt
PyCrypto